ФИКС тестов

+++ ewm-main-svc/src/main/java/ru/practicum/ewm/user/service/UserServiceImp.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.ewm.exception.ConflictException;
+import ru.practicum.ewm.exception.NotFoundException;
 import ru.practicum.ewm.user.dto.UserCreateDto;
 import ru.practicum.ewm.user.dto.UserMapper;
 import ru.practicum.ewm.user.model.User;
@@ -37,6 +38,9 @@ public class UserServiceImp implements UserService {
 
     @Override
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new NotFoundException("Пользователь с id " + id + " не существует");
+        }
         userRepository.deleteById(id);
         log.info("Пользователь с id: {} успешно удален", id);
     }
